requirements.txt
splunk-sdk==1.7.3
requests>=2.28.0
pyyaml>=6.0
click>=8.1.0
jinja2>=3.1.0
tabulate>=0.9.0
colorama>=0.4.6
cryptography>=41.0.0
python-dateutil>=2.8.2